accept floats as numerical inputs

+++ assignment_01.py
@@ -7,9 +7,9 @@
 HIGH_SCORE = 95
  
 # Get the test scores.
-test1 = int(input('Enter the score for test 1: '))
-test2 = int(input('Enter the score for test 2: '))
-test3 = int(input('Enter the score for test 3: '))
+test1 = float(input('Enter the score for test 1: '))
+test2 = float(input('Enter the score for test 2: '))
+test3 = float(input('Enter the score for test 3: '))
 # Calculate the average test score.
 average = (test1 + test2 + test3) / 3
 # Print the average.
@@ -23,13 +23,13 @@ print('That is a great average!')
 
 #Q2
 #The area of a rectangle is the rectangle’s length times its width. Write a program that asks for the length and width of two rectangles and prints to the user the area of both rectangles. 
-length_1 = int(input('Enter the length for rectangle 1: '))
-width_1 = int(input('Enter the width for rectangle 1: '))
+length_1 = float(input('Enter the length for rectangle 1: '))
+width_1 = float(input('Enter the width for rectangle 1: '))
 
 print("The area of rectangle 1 is: " + str(length_1*width_1))
 
-length_2 = int(input('Enter the length for rectangle 2: '))
-width_2 = int(input('Enter the width for rectangle 2: '))
+length_2 = float(input('Enter the length for rectangle 2: '))
+width_2 = float(input('Enter the width for rectangle 2: '))
 
 print("The area of rectangle 1 is: " + str(length_2*width_2))
 
@@ -41,5 +41,5 @@ print("The area of rectangle 1 is: " + str(length_2*width_2))
 # "Happy birthday, name!  You are age years old today!"
 
 name = input('Please enter your first name: ')
-age = int(input('Please enter your age: '))
+age = float(input('Please enter your age: '))
 print('Happy birthday ' + name + '! You are ' + str(age) + ' years old today!')
